Split prescribers tests in 3 files
